pyproject.toml
[project]
name = "web-remake"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivy-Chen199",email = "yuxin.chen@helsinki.fi"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "flask-testing (>=0.8.1,<0.9.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"
pytest = "^8.3.5"
pytest-flask = "^1.3.0"

